Max bandwidth setting is much more accurate.
It actually didn't work before, but may have seemed like it did on wired connections because it had a hard-limit of 60Mb/s.
It now takes into account audio overhead and halves video bitrate since we seem to send up to 1 second of video in advance, or at least that's my guess as to why the network bandwidth is often close to double the video bitrate.