Removing skip() override introduced in 562fdb8.
The method interfered with the correctness of isCompatible() and its
backwards compatibility with ps3compatible().

On top of that the comment seemed to want to force transcoding, where
in reality returning true ended up forcing streaming. To remove this
paradox and to restore control to each individual renderer
configuration the method has been removed, bringing the code in line
with all other formats.

From git 7675911fcb177dcbfd3eb38398dd0197fb912973